Use `pull_request_target` event
GitHub security policy disallows sharing secrets with forks in public
repositories [1]. However, we need pull requests to push temporary
Docker images to a registry, which requires access to secrets unless we
use GitHub Packages.

To work around this, this commit switches to the `pull_request_target`
event, which works on the repository and commit that the pull request is
based on, but manually checks out and uses the code from the pull
request. This is a mild security hazard but should be OK, since
workflows on pull requests from new contributors need to be approved
anyways.

 [1]: https://securitylab.github.com/research/github-actions-preventing-pwn-requests/