Move LaTeX section about finalizing and freezing cache
Previously, the section was located between the section about generating
plain TeX option macros and key=values, and the sections about
generating key=values for renderers and renderer prototypes, where it
did not belong. I moved it to the beginning of the section about LaTeX
options and gave it a heading that corresponds to a related section in
the section about plain TeX options.